+++ src/main/java/com/mitchej123/hodgepodge/mixins/early/fml/MixinASMDataTable.java
@@ -1,4 +1,4 @@
-package com.mitchej123.hodgepodge.mixins.early.forge;
+package com.mitchej123.hodgepodge.mixins.early.fml;
 
 import java.io.File;
 import java.util.HashMap;
