+++ src/main/java/com/mitchej123/hodgepodge/mixins/early/fml/MixinFMLProxyPacket.java
@@ -1,4 +1,4 @@
-package com.mitchej123.hodgepodge.mixins.early.forge;
+package com.mitchej123.hodgepodge.mixins.early.fml;
 
 import net.minecraft.network.Packet;
