+++ src/main/java/com/mitchej123/hodgepodge/mixins/early/fml/MixinNetworkDispatcher.java
@@ -1,4 +1,4 @@
-package com.mitchej123.hodgepodge.mixins.early.forge;
+package com.mitchej123.hodgepodge.mixins.early.fml;
 
 import java.lang.ref.WeakReference;
